Feat/execution report (#42)
* Add tracking numbers to market events and a unique ID to client events

* Update random usage of different integer sizes to use a consistent typedef

* Add side to generated events

* Add function to convert ClientEvents into execution reports

* Plumb ClientEvent pipeline to FIXEngine for sending

+++ crates/pyclient/src/fix/client.rs
@@ -414,12 +414,12 @@ mod tests {
         std::thread::sleep(Duration::from_secs(2));
 
         let order = Order {
-            order_id: 1 as u64,
+            order_id: 1,
             side: OrderSide::Bid,
-            timestamp: 5 as u64,
+            timestamp: 5,
             kind: OrderType::Limit {
-                qty: 10 as u64,
-                price: 666 as u64,
+                qty: 10,
+                price: 666,
             },
         };
